+++ src/Wrapper/index.tsx
@@ -1,12 +1,10 @@
 'use client'
 
-import React, { lazy, FC, Children, isValidElement, cloneElement } from 'react'
+import React, { lazy, FC } from 'react'
 import { Box, Grid } from '@mui/material'
 
 // Types
-import type { ReactElement } from 'react'
 import type { SpinoramaWrapperProps } from './index.types'
-import type { SpinoramaItemProps } from '../Item/index.types'
 
 // Components
 const SpinoramaItem = lazy(() => import('../Item'))
@@ -31,13 +29,7 @@ const SpinoramaWrapper: FC<SpinoramaWrapperProps> = (props: SpinoramaWrapperProp
 				minWidth={1}
 				height={1}
 			>
-				{Children.map(children, (child, index) => {
-					if (isValidElement(child)) {
-						// Clone element
-						if (child.type.toString().indexOf('spinorama-item') > -1) return cloneElement(child as ReactElement<SpinoramaItemProps>)
-						else return <SpinoramaItem key={index}>{child}</SpinoramaItem>
-					} else return child
-				})}
+				{children}
 			</Grid>
 		</Box>
 	)